+++ CHANGES.md
@@ -4,11 +4,14 @@
 
 ### Bug fixes
 
--   Fix error in `write_dataframe` with `use_arrow=False` when writing a column with
-    datetimes without any time zone offsets or with a mix of offsets and no offsets
-    (#634).
+-   Fix error in `write_dataframe` with `use_arrow=False` when writing an object-type
+    column with datetimes without any time zone offsets or with a mix of offsets and
+    no offsets (#634).
+-   Fix writing non-string object columns with arrow (#630).
 -   Fix writing empty string category columns with arrow fails (#621).
+-   Fix Time type columns being skipped with `use_arrow=False` (#617).
 -   Fix overwriting a corrupt fileGDB directory (#600).
+-   Fix attribute data being incorrectly written with KML driver (#650).
 
 ## 0.12.1 (2025-11-28)